separate configuration of services out from application module
